# new/protocol/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Invity",
    "description": "Volvo API.",
    "version": "0.1"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/project": {
      "get": {
        "summary": "remove last response",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean",
                  "description": "is user a mentor of the project"
                },
                "leader": {
                  "type": "boolean",
                  "description": "is user a leader of the project"
                },
                "member": {
                  "type": "boolean",
                  "description": "is user a member of the project"
                },
                "favourite": {
                  "type": "boolean",
                  "description": "is user's favourite project"
                },
                "recruitment": {
                  "type": "boolean",
                  "description": "is recruitment open"
                },
                "invited": {
                  "type": "boolean",
                  "description": "has been user invited to the project"
                },
                "applied": {
                  "type": "boolean",
                  "description": "has user applied to the project"
                },
                "project": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "depiction": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string",
                      "format": "path"
                    }
                  }
                },
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "skill": {
                        "type": "string",
                        "description": "title of skill connected with role"
                      },
                      "members": {
                        "type": "array",
                        "description": "all project members involved in role",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "avatar": {
                              "type": "string",
                              "format": "path"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      },
                      "link": {
                        "type": "string",
                        "format": "url"
                      }
                    }
                  }
                },
                "closure": {
                  "type": "object",
                  "properties": {
                    "closed": {
                      "type": "boolean",
                      "description": "is project closed"
                    },
                    "state": {
                      "type": "boolean",
                      "description": "was closure request sent by the leader"
                    },
                    "explanation": {
                      "type": "string",
                      "description": "the reason why the leader want to close the project"
                    }
                  }
                },
                "response": {
                  "type": "object",
                  "description": "message thet will be shown when user opens a project page; often reffered as 'send message'",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project"
        ]
      }
    },
...
    "/project/close": {
      "post": {
        "summary": "set up a proper notification; close project",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/close?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "agreement": {
                  "type": "string"
                },
                "success": {
                  "type": "string"
                },
                "summary": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/project/closure_request": {
      "post": {
        "summary": "set up a proper notification; send closing request to the mentor",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/closure_request?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "explanation": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/project/report": {
      "post": {
        "summary": "report project",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/report?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "explanation": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/project/description": {
      "post": {
        "summary": "update project description",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/description?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
    "/project/media": {
      "post": {
        "summary": "update project media",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /project/media?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "media": {
                  "type": "array"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project"
        ]
      }
    },
...
    "/recruitment": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "skill": {
                        "type": "string",
                        "format": "skill_code"
                      },
                      "recruitment": {
                        "type": "boolean",
                        "description": "is recruitment open"
                      },
                      "recommended": {
                        "type": "integer",
                        "description": "how meny users have seen this role in theirs recomendations"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      }
                    }
                  }
                },
                "applications": {
                  "type": "array",
                  "description": "all applications for the project",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "surname": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string",
                        "format": "user_id"
                      },
                      "member": {
                        "type": "boolean",
                        "description": "is user a member of the project"
                      },
                      "roles": {
                        "type": "array",
                        "description": "titles of roles they applied",
                        "items": {
                          "type": "string"
                        }
                      },
                      "accepted": {
                        "type": "array",
                        "description": "was user accepted for the according role",
                        "items": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "skills": {
                  "type": "array",
                  "description": "all available skills",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "role_title": {
                        "type": "string",
                        "description": "default title of the role (based on the skill title)"
                      },
                      "role_description": {
                        "type": "string",
                        "description": "default description of the role (based on the skill title)"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      },
                      "hash": {
                        "type": "string",
                        "format": "skill_code"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
...
    "/reports": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean"
                },
                "leader": {
                  "type": "boolean"
                },
                "closed": {
                  "type": "boolean"
                },
                "reports": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "description": "project reports",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "answers": {
                        "type": "array",
                        "description": "answers to 4 questions in report form",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "responses": {
                  "description": "responses from mentor",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "date": {
                        "type": "string",
                        "format": "date"
                      },
                      "text": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project",
          "project/reports"
        ]
      }
    },
...
    "/team": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean"
                },
                "leader": {
                  "type": "boolean"
                },
                "recruitment": {
                  "type": "boolean"
                },
                "members": {
                  "type": "array",
                  "items": {
                    "description": "all project members",
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "surname": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "description": "aka bio"
                      },
                      "avatar": {
                        "type": "string",
                        "format": "path"
                      },
                      "id": {
                        "type": "string",
                        "format": "user_id"
                      },
                      "roles": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "title": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "appreciated": {
                              "type": "boolean",
                              "description": "was the member appreciated - has level 3 oh the skill"
                            },
                            "image": {
                              "type": "string",
                              "format": "path"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "invitations": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "user name or email address"
                      },
                      "surname": {
                        "type": "string",
                        "description": "surname or empty string"
                      },
                      "date": {
                        "type": "string",
                        "format": "date",
                        "description": "date of the invitation sending"
                      },
                      "image": {
                        "type": "string",
                        "format": "path"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project",
          "project/team"
        ]
      }
    },
...
    "/team/report": {
      "post": {
        "summary": "report member",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /team/report",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "project": {
                  "type": "string"
                },
                "explanation": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/team"
        ]
      }
    },
...
  },
...
}